Refactor tests and add service layer coverage
